pyproject.toml
[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "aif_workflow_helper"
namespace = true

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[project]
name = "aif-workflow-helper"
version = "0.3.1"
description = "Azure AI Foundry dev workflow helpers"
authors = [
  { name = "Pete Roden", email = "pete.roden@live.com" }
]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["azure", "ai", "agents", "workflow", "automation"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

# Core runtime dependencies
dependencies = [
  "azure-ai-projects",
  "azure-ai-agents==1.2.0b5",
  "agent-framework==1.0.0b251007",
  "agent-framework-azure-ai==1.0.0b251007",
  "azure-identity",
  "pyyaml",
  "python-frontmatter",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
  "pytest-cov>=4.0.0",
  "pytest-mock>=3.10.0",
]

[project.urls]
Homepage = "https://github.com/peteroden/aif-workflow-helper"
Source = "https://github.com/peteroden/aif-workflow-helper"

# Console script entry points
[project.scripts]
aif-workflow-helper = "aif_workflow_helper.cli.main:main"
